extracted properties to HasPropertyTrait
