+++ src/Types/StructuralTypes/RelationshipType.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace WikibaseSolutions\CypherDSL\Types\StructuralTypes;
+
+interface RelationshipType
+{
+
+}
